+++ python/manr/common/definitions/state_definitions.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from enum import StrEnum
-
 import attr
 import numpy as np
 from manr.common.custom_types import (
@@ -10,18 +8,7 @@ from manr.common.custom_types import (
     JointPositionsVector,
     JointVelocitiesVector,
 )
-
-
-class ManipulatorGripperStatus(StrEnum):
-    """
-    Enum to represent the status of a manipulator's gripper.
-    """
-
-    OPENING = "opening"
-    CLOSING = "closing"
-    NEUTRAL = "neutral"
-    OPEN = "open"
-    CLOSED = "closed"
+from manr.common.definitions.control_definitions import ManipulatorGripperStatus
 
 
 @attr.frozen
